% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{add_dummy_coverage_to_bedgraphs}
\alias{add_dummy_coverage_to_bedgraphs}
\title{Add dummy coverage values columns to bedGraphs so that they can be used with methrix::read_bedgraphs()}
\usage{
add_dummy_coverage_to_bedgraphs(
  bedgraphs,
  output_directory = "dummy_coverage_bedgraphs",
  ncores = 1,
  tmpdir = tempdir()
)
}
\arguments{
\item{bedgraphs}{A list of paths to bedGraph files}

\item{output_directory}{A directory in which to save the modified bedGraph files. Directory is created if it doesn't exist. Default name is "dummy_coverage_bedgraphs".
Can overwrite the original bedGraphs if output_directory is the directory in which they are located.}

\item{ncores}{Number of cores to use. Default is 1.}
}
\value{
Invisibly returns TRUE if all bedGraphs converted without errors.
}
\description{
Each CpG site is given a coverage value of 50.
}
