% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_meth_site_transcript_partial_cors.R
\name{adjust_covariates}
\alias{adjust_covariates}
\title{Correct methylation values or gene expression values for covariates}
\usage{
adjust_covariates(values, covariates)
}
\arguments{
\item{values}{A table with methylation values or expression values to correct}

\item{covariates}{Either a data.frame with covariates to correct values for or a design
matrix constructed with model.matrix(). Row names should match column names of values.}
}
\value{
A matrix with the same dimensions of values containing the methylation or expression values
corrected for the covariates.
}
\description{
Corrects for covariates by regressing out their effects using linear models.
}
