% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_annotation.R
\name{annotate_granges}
\alias{annotate_granges}
\title{Annotate GRanges}
\usage{
annotate_granges(
  genomic_regions,
  annotation_ranges,
  annotation_column = "region_type",
  ignore.strand = TRUE,
  overlap_measure = "absolute"
)
}
\arguments{
\item{genomic_regions}{A GRanges object to be annotated}

\item{annotation_ranges}{A GRanges object with regions for different features e.g. introns, exons, enhancers.}

\item{annotation_column}{Name of the metadata column of annotation_ranges indicating the feature group that regions belong to. Default is "region_type".}

\item{ignore.strand}{A logical value indicating whether strand should be ignored when calculating intersections. Default is TRUE.}

\item{overlap_measure}{One of "absolute", "proportion" or "jaccard" indicating whether to calculate
the absolute size of the intersection in base pairs, the proportion base paris of gr1 overlapping gr2
or the Jaccard index of the intersection in terms of base pairs. Default value is "absolute".}
}
\value{
A numeric vector with the overlap measure for genomic_regions with each type of region in annotation_ranges
}
\description{
Annotate GRanges
}
