% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{annotate_meth_site_plot}
\alias{annotate_meth_site_plot}
\title{Create a plot with genomic annotation for a plot returned by plot_meth_site_values()}
\usage{
annotate_meth_site_plot(
  meth_site_plot,
  annotation_gr,
  reference_region = NULL,
  region_class_colours = NULL,
  annotation_line_size = 5,
  annotation_plot_height = 0.5,
  keep_meth_site_plot_legend = FALSE,
  annotation_plot_only = FALSE
)
}
\arguments{
\item{meth_site_plot}{A plot of methylation site values (generally methylation level or correlation of methylation with transcription) around a TSS}

\item{annotation_gr}{A GRanges object giving the locations of different classes of regions.
There must be a metadata column named region_type giving the class of each region.
If this is a factor, the levels will be used to order the region classes in the plot.}

\item{reference_region}{An optional GRanges object with a single range.
If provided, the x-axis will show the distance of methylation sites to the start of this region with methylation sites upstream
relative to the reference_region shown first. If not, the x-axis will show the start site coordinate of the methylation site.}

\item{region_class_colours}{An optional named vector of colours to use with different region classes. Names of vector match colours to region classes.}

\item{annotation_line_size}{Linewidth for annotation plot. Default is 5.}

\item{annotation_plot_height}{A value giving the proportion of the height of the plot devoted to the annotation. Default is 0.5.}

\item{keep_meth_site_plot_legend}{A logical value indicating whether to retain the legend of meth_site_plot, if it has one. Default value is FALSE.}

\item{annotation_plot_only}{A logical value indicating whether to return only the annotation plot. Default is to combine meth_site_plot with the annotation.}
}
\value{
A ggplot object
}
\description{
Can combine the meth site values plot and genomic annotation together into a single plot or return the annotation plot separately.
}
\examples{
# Get CpG islands from UCSC
ucsc_query = rtracklayer::ucscTableQuery("hg38", table = "cpgIslandExt")
cpg_islands = rtracklayer::track(ucsc_query) 
cpg_islands$region_type = "CpG Island"

# Load plot with CpG methylation correlation values for TUBB6
data("tubb6_correlation_plot", package = "methodical")

# Add positions of CpG islands to tubb6_correlation_plot
methodical::annotate_meth_site_plot(tubb6_correlation_plot, cpg_islands, annotation_plot_height = 0.3)

}
