% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{annotateMethSitePlot}
\alias{annotateMethSitePlot}
\title{Create a plot with genomic annotation for a plot returned by plotMethSiteValues()}
\usage{
annotateMethSitePlot(
  meth_site_plot,
  annotation_gr,
  reference_tss = NULL,
  region_class_colours = NULL,
  annotation_line_size = 5,
  annotation_plot_height = 0.5,
  keep_meth_site_plot_legend = FALSE,
  annotation_plot_only = FALSE
)
}
\arguments{
\item{meth_site_plot}{A plot of Value (generally methylation level or correlation of methylation with transcription) around a TSS}

\item{annotation_gr}{A GRanges object giving the locations of different classes of regions.
There must be a metadata column named region_type giving the class of each region.
If this is a factor, the levels will be used to order the region classes in the plot.}

\item{reference_tss}{A logical value indicating whether to show distances on the X-axis
relative to the TSS stored as an attribute \code{tss_range} of meth_site_plot.
Alternatively, can provide a GRanges object with a single range for such a TSS site.
In either case, will show the distance of methylation sites to the start of this region with methylation sites upstream
relative to the reference_tss shown first.
If FALSE (the default), the x-axis will instead show the start site coordinate of the methylation site.
relative to the reference_tss shown first. If not, the x-axis will show the start site coordinate of the methylation site.}

\item{region_class_colours}{An optional named vector of colours to use with different region classes. Names of vector match colours to region classes.}

\item{annotation_line_size}{Linewidth for annotation plot. Default is 5.}

\item{annotation_plot_height}{A value giving the proportion of the height of the plot devoted to the annotation. Default is 0.5.}

\item{keep_meth_site_plot_legend}{A logical value indicating whether to retain the legend of meth_site_plot, if it has one. Default value is FALSE.}

\item{annotation_plot_only}{A logical value indicating whether to return only the annotation plot. Default is to combine meth_site_plot with the annotation.}
}
\value{
A ggplot object
}
\description{
Can combine the meth site values plot and genomic annotation together into a single plot or return the annotation plot separately.
}
\examples{
# Get CpG islands from UCSC
cpg_island_annotation <- annotatr::build_annotations(genome = "hg38", annotations = "hg38_cpgs")
cpg_island_annotation$region_type <- cpg_island_annotation$type

# Load plot with CpG methylation correlation values for TUBB6
data("tubb6_correlation_plot", package = "methodical")

# Add positions of CpG islands to tubb6_correlation_plot
methodical::annotateMethSitePlot(tubb6_correlation_plot, annotation_gr = cpg_island_annotation, annotation_plot_height = 0.3)

}
