% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_documentation.R
\docType{data}
\name{arabidopsis_chr4}
\alias{arabidopsis_chr4}
\title{arabidopsis_chr4}
\format{
A GRanges object.
}
\usage{
arabidopsis_chr4
}
\description{
A DNAStringSet for chr4 from BSgenome.Athaliana.TAIR.TAIR9
}
\keyword{datasets}
