% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_documentation.R
\docType{data}
\name{chr11_subset_hg38_cpgs}
\alias{chr11_subset_hg38_cpgs}
\title{chr11_subset_hg38_cpgs}
\format{
A GRanges object.
}
\usage{
chr11_subset_hg38_cpgs
}
\description{
All the CpG sites on both strands within the region chr11:67578812-67588812
}
\keyword{datasets}
