% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_documentation.R
\docType{data}
\name{chr18_subset_hg38}
\alias{chr18_subset_hg38}
\title{chr18_subset_hg38}
\format{
A DNAStringSet object.
}
\usage{
chr18_subset_hg38
}
\description{
A DNAStringSet for first 1,000,000 bp from chr18 from BSgenome.Hsapiens.UCSC.hg38
}
\keyword{datasets}
