% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_documentation.R
\docType{data}
\name{chr1_subset_hg38}
\alias{chr1_subset_hg38}
\title{chr1_subset_hg38}
\format{
A DNAStringSet object.
}
\usage{
chr1_subset_hg38
}
\description{
A DNAStringSet for first 1,000,000 bp from chr1 from BSgenome.Hsapiens.UCSC.hg38
}
\keyword{datasets}
