% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_documentation.R
\docType{data}
\name{chr4_subset_a_thal}
\alias{chr4_subset_a_thal}
\title{chr4_subset_a_thal}
\format{
A DNAStringSet object.
}
\usage{
chr4_subset_a_thal
}
\description{
A DNAStringSet for first 1,000,000 bp from chr4 from BSgenome.Athaliana.TAIR.TAIR9
}
\keyword{datasets}
