% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcript_cpg_correlations2.R
\name{combine_transcription_cpg_correlations}
\alias{combine_transcription_cpg_correlations}
\title{Created a data.frame combining the results from the output of transcription_cpg_correlations()}
\usage{
combine_transcription_cpg_correlations(
  transcription_cpg_correlations_list,
  correct_p_values = T,
  p_adjust_method = "BH"
)
}
\arguments{
\item{transcription_cpg_correlations_list}{A list created by transcription_cpg_correlations()}

\item{correct_p_values}{Logical value indicating whether p-values in the final table should be corrected, Default is TRUE.}

\item{p_adjust_method}{Method used to adjust p-values if correct_p_values is specified. Same as the methods from p.adjust.methods. Default is Benjamini-Hochberg.}
}
\description{
Created a data.frame combining the results from the output of transcription_cpg_correlations()
}
