% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meth_rse_functions.R
\name{convert_rse_to_bsseq}
\alias{convert_rse_to_bsseq}
\title{Convert a methylation RSE to a BSseq object}
\usage{
convert_rse_to_bsseq(meth_rse, proportion_assay = 1, coverage_assay = 2)
}
\arguments{
\item{meth_rse}{A RangedSummarizedExperiment with methylation values.}

\item{proportion_assay}{The assay of meth_rse which corresponds to the proportion of methylated reads.
Can be either a numeric index or the name of the assay. Default is the first assay.}

\item{coverage_assay}{The assay of meth_rse which corresponds to the coverage (the total number of reads).
Can be either a numeric index or the name of the assay. Default is the second assay.}
}
\value{
A RangedSummarized experiment identical to meth_rse with two additional assays added: one for methylated reads and another for unmethylated reads.
}
\description{
Convert a methylation RSE to a BSseq object
}
