% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmr_functions.R
\name{correct_correlation_pvalues}
\alias{correct_correlation_pvalues}
\title{Correct p-values for a list of methylation-transcription correlations results}
\usage{
correct_correlation_pvalues(correlation_list, p_adjust_method = "fdr")
}
\arguments{
\item{correlation_list}{A list of data.frames with correlation values between methylation sites
and a transcript as returned by calculateMethSiteTranscriptCors.}

\item{p_adjust_method}{The method to use for p-value adjustment. Should be one of the methods in p.adjust.methods.
Default is "fdr".}
}
\value{
A list identical to correlation_list except with p-values corrected using the indicated method.
}
\description{
Correct p-values for a list of methylation-transcription correlations results
}
