% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/granges_manipulation.R
\name{count_covered_bases}
\alias{count_covered_bases}
\title{Calculate the number of unique bases covered by all regions in a GRanges object}
\usage{
count_covered_bases(gr)
}
\arguments{
\item{gr}{A GRanges object}
}
\value{
An numeric value
}
\description{
Calculate the number of unique bases covered by all regions in a GRanges object
}
