% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{cpg_distances_to_region}
\alias{cpg_distances_to_region}
\title{Find distances between start of a region and CpG sites}
\usage{
cpg_distances_to_region(query_region, cpg_names)
}
\arguments{
\item{query_region}{A GRanges object with a single range for one query region.}

\item{cpg_names}{Names of CpG sites in the form sequence:start where start is 1-based e.g. chr1:10469 for the first CpG site on chr1 from hg38.}
}
\value{
None
}
\description{
Find distances between a query region and CpG sites, giving a negative distance if CpGs are upstream of the query region and positive values if they are downstream.
Adds 1 to the distance found using CpGs GenomicRanges::distance() so that CpGs touching a query_region have a distance of 1 and
CpGs located within the region are assigned a distance of 0 bases from the regions.
}
