% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{cpg_feature_cor}
\alias{cpg_feature_cor}
\title{Calculate correlation between methylation of CpG sites and a feature}
\usage{
cpg_feature_cor(cpg_table, feature, method = "pearson")
}
\arguments{
\item{cpg_table}{A data.frame with methylation values for CpG sites where CpG names are rownames and column names are samples.}

\item{feature}{A numeric vector for a feature of interest with sample names as observation names.}

\item{method}{A character string indicating which correlation coefficient is to be computed. Identical to methods from cor().}
}
\value{
A data.frame with the correlation values for all CpGs in table2 with feature
}
\description{
Calculate correlation between methylation of CpG sites and a feature
}
