% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylation_summary.R
\name{cpg_methylation_summary}
\alias{cpg_methylation_summary}
\title{Summarize methylation values for CpG regions surrounding genomic regions. MAYBE LIMIT THE FUNCTIONS THAT CAN BE USED WITH SUMMARY_FUNCTION.
MAYBE CAN SPEED UP SAMPLE_GROUP2 DIFFERENCE.}
\usage{
cpg_methylation_summary(
  methrix_object,
  samples_group1 = NULL,
  samples_group2 = NULL,
  genomic_regions,
  expand_upstream = 0,
  expand_downstream = 0,
  summary_function = rowMeans,
  summary_column_name = "methylation_summary",
  add_distance_to_region = T,
  ncores = 1
)
}
\arguments{
\item{methrix_object}{A methrix object. Names of samples must match those in transcript_expression_table unless samples_group1 is provided.}

\item{samples_group1}{Sample names to subset methrix object and transcript_expression_table. Provided samples must be found in both methrix_object and transcript_expression_table.
Default is to use all samples in methrix_object and transcript_expression_table.}

\item{samples_group2}{Sample names to subset methrix object and transcript_expression_table. Provided samples must be found in both methrix_object and transcript_expression_table.
Default is to use all samples in methrix_object and transcript_expression_table.}

\item{genomic_regions}{Genomic regions to calculate mean CpG methylation for. There should be one region for each row in transcript_expression_table.
Must have a metadata column called transcript_id that is identical to the row.names of transcript_expression_table.}

\item{expand_upstream}{Number of bases to add upstream of each region in genomic regions. Must be numeric vector of length 1 or equal to the length of genomic_regions. Default is 0.}

\item{expand_downstream}{Number of bases to add downstream of each region in genomic regions. Must be numeric vector of length 1 or equal to the length of genomic_regions. Default is 0.}

\item{summary_function}{A summary function to apply to the methylation values of each CpG site. Default is rowMeans().}

\item{summary_column_name}{The name to give the column with the output of the summary function. Default is "methylation_summary".}

\item{add_distance_to_region}{A logical value indicating whether to add the distance of CpG sites to the center of the associated genomic region. Default value is TRUE.
Setting to FALSE will roughly half the total running time.}

\item{ncores}{Number of cores to use. Default is 1.}
}
\value{
A list of data.frames with a summary of the methylation values of CpG sites surrounding a specified genomic region.
Distance of the CpG upstream or downstream to the center of the region is also provided.
}
\description{
Summarize methylation values for CpG regions surrounding genomic regions. MAYBE LIMIT THE FUNCTIONS THAT CAN BE USED WITH SUMMARY_FUNCTION.
MAYBE CAN SPEED UP SAMPLE_GROUP2 DIFFERENCE.
}
