% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{cpg_window_summary}
\alias{cpg_window_summary}
\title{Summarize CpG methylation across genomic windows}
\usage{
cpg_window_summary(
  cpg_values,
  window_size = 1000,
  summary_function = mean,
  min_cpgs_per_window = 1
)
}
\arguments{
\item{cpg_values}{A data.frame with the methylation values of CpGs for different samples such as returned by extract_cpg_values_from_methrix.}

\item{window_size}{The size of the windows in which to bin CpGs. Default is 1000.}

\item{summary_function}{Summary function to apply to CpG methylation values in windows. Default is mean. Can be any function which returns a single value from a numeric vector.}

\item{min_cpgs_per_window}{The minimum number of non-missing CpG methylation values needed to assign a value to a window. Windows with less than this value are given a value of NA. Default is 1.}
}
\value{
A data.frame with a summary of CpG methylation across genomic windows for each sample. Row names are the names of the genomic windows.
}
\description{
Summarize CpG methylation across genomic windows
}
