% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_meth_rse_accessory_functions.R
\name{.check_input_files}
\alias{.check_input_files}
\title{Check input files have the correct number of columns specified and that columns seem to be of correct type}
\usage{
.check_input_files(input_files, meth_files_columns)
}
\arguments{
\item{input_files}{A vector of input filepaths.}

\item{meth_files_columns}{A list specifying the columns in meth_files.}
}
\description{
Check input files have the correct number of columns specified and that columns seem to be of correct type
}
