% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_meth_rse_accessory_functions.R
\name{.collapse_strands}
\alias{.collapse_strands}
\title{Combine values for stranded data}
\usage{
.collapse_strands(meth_df, sequence_context)
}
\arguments{
\item{meth_df}{A data.frame with methylation data.}

\item{sequence_context}{A single character string or DNAString with the sequence context of the methylation sites e.g. CG or CHG.}
}
\description{
Combine values for stranded data
}
