% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_meth_rse_accessory_functions.R
\name{.create_meth_rse_from_hdf5}
\alias{.create_meth_rse_from_hdf5}
\title{Create a RangedSummarizedExperiment for methylation values already deposited in HDF5}
\usage{
.create_meth_rse_from_hdf5(
  hdf5_filepath,
  hdf5_dir,
  meth_sites,
  sample_metadata
)
}
\arguments{
\item{hdf5_filepath}{Path to HDF5 file}

\item{hdf5_dir}{The path to the HDF5 directory.}

\item{meth_sites}{A sorted GRanges object with the locations of the methylation sites of interest.}

\item{sample_metadata}{A data.frame with sample metadata}
}
\value{
A RangedSummarizedExperiment with methylation values
}
\description{
Create a RangedSummarizedExperiment for methylation values already deposited in HDF5
}
