% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_meth_site_transcript_cors_update2.R
\name{.expand_transcript_ranges}
\alias{.expand_transcript_ranges}
\title{Add regions upstream of TSS and downstream of TES to a GRangesList for transcripts}
\usage{
.expand_transcript_ranges(grl, expand_upstream = 0, expand_downstream = 0)
}
\arguments{
\item{grl}{A GRangesList object with ranges for exons and introns of each transcript}

\item{expand_upstream}{Number of bases to add upstream of TSS each transcript. Must be numeric vector of length 1 or equal to the length of tss_gr.}

\item{expand_downstream}{Number of bases to add downstream of TES of each transcript. Must be numeric vector of length 1 or equal to the length of tss_gr.}
}
\value{
A GRangesList object
}
\description{
Add regions upstream of TSS and downstream of TES to a GRangesList for transcripts
}
