% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_meth_rse_accessory_functions.R
\name{.set_meth_df_columns}
\alias{.set_meth_df_columns}
\title{Process a data.frame with methylation data so that it contains the correct columns}
\usage{
.set_meth_df_columns(meth_df, zero_based)
}
\arguments{
\item{meth_df}{A data.frame with methylation data.}

\item{zero_based}{TRUE or FALSE indicating if files are zero-based.}
}
\description{
Process a data.frame with methylation data so that it contains the correct columns
}
