% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_meth_rse_accessory_functions.R
\name{.split_bedgraph}
\alias{.split_bedgraph}
\title{Split data from a single bedgraph file into chunks}
\usage{
.split_bedgraph(bg_file, column, file_count, parameters)
}
\arguments{
\item{bg_file}{Path to a bedgraph file.}

\item{column}{The current grid column being processed.}

\item{file_count}{The number of the current file being processed.}

\item{parameters}{A list of parameters for processing the bedgraph.}
}
\value{
Invisibly returns NULL.
}
\description{
Split data from a single bedgraph file into chunks
}
