% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_meth_rse_accessory_functions.R
\name{.split_bedgraphs_into_chunks}
\alias{.split_bedgraphs_into_chunks}
\title{Split data from bedGraph files into chunks}
\usage{
.split_bedgraphs_into_chunks(
  bedgraphs,
  seqnames_column,
  start_column,
  end_column,
  value_column,
  file_grid_columns,
  meth_sites,
  meth_site_groups,
  temp_chunk_dirs,
  zero_based,
  normalization_factor,
  decimal_places,
  BPPARAM
)
}
\arguments{
\item{bedgraphs}{Paths to bedgraph files.}

\item{seqnames_column}{The column number in bedgraphs which corresponds to the sequence names.}

\item{start_column}{The column number in bedgraphs which corresponds to the start positions.}

\item{end_column}{The column number in bedgraphs which corresponds to the end positions.}

\item{value_column}{The column number in bedgraphs which corresponds to the methylation values.}

\item{file_grid_columns}{The grid column number for each file.}

\item{meth_sites}{A GRanges object with the locations of the methylation sites of interest.}

\item{meth_site_groups}{A list with the indices of the methylation sites in each group.}

\item{temp_chunk_dirs}{A vector giving the temporary directory associated with each chunk.}

\item{zero_based}{A logical value indicating if files are zero-based.}

\item{normalization_factor}{An optional numerical value to divide methylation values by to convert them to fractions e.g. 100 if they are percentages.
Default is not to leave values as they are in the input files.}

\item{decimal_places}{Integer indicating the number of decimal places to round beta values to.}

\item{BPPARAM}{A BiocParallelParam object.}
}
\value{
A data.table with the methylation sites sorted by seqnames and start.
}
\description{
Split data from bedGraph files into chunks
}
