% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_meth_rse_accessory_functions.R
\name{.split_meth_array_files_into_chunks}
\alias{.split_meth_array_files_into_chunks}
\title{Split data from methylation array files into chunks}
\usage{
.split_meth_array_files_into_chunks(
  array_files,
  probe_name_column,
  beta_value_column,
  file_grid_columns,
  probe_ranges,
  probe_groups,
  temp_chunk_dirs,
  decimal_places,
  BPPARAM
)
}
\arguments{
\item{array_files}{Paths to methylation array files.}

\item{probe_name_column}{The column number in array_files which corresponds to the name of the probes. Default is 1st column.}

\item{beta_value_column}{The column number in array_files which corresponds to the beta values. Default is 2nd column.}

\item{file_grid_columns}{The grid column number for each file.}

\item{probe_ranges}{A GRanges object giving the genomic locations of probes where each region corresponds to a separate probe.}

\item{probe_groups}{A list with the indices of the probes in each group.}

\item{temp_chunk_dirs}{A vector giving the temporary directory associated with each chunk.}

\item{decimal_places}{Integer indicating the number of decimal places to round beta values to.}

\item{BPPARAM}{A BiocParallelParam object.}
}
\value{
A data.table with the probe sites sorted by seqnames, start and probe name.
}
\description{
Split data from methylation array files into chunks
}
