% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_meth_rse_accessory_functions.R
\name{.split_meth_file}
\alias{.split_meth_file}
\title{Split data from a single input methylation file into chunks}
\usage{
.split_meth_file(
  meth_file,
  meth_files_columns,
  grid_column,
  file_count,
  parameters
)
}
\arguments{
\item{meth_file}{Path to an input methylation file.}

\item{meth_files_columns}{A list specifying the columns in meth_files.}

\item{grid_column}{The current grid column being processed.}

\item{file_count}{The number of the current file being processed.}

\item{parameters}{A list of parameters for processing meth_file.}
}
\value{
Invisibly returns NULL.
}
\description{
Split data from a single input methylation file into chunks
}
