% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_meth_rse_accessory_functions.R
\name{.split_meth_files_into_chunks}
\alias{.split_meth_files_into_chunks}
\title{Split data from input methylation files into chunks}
\usage{
.split_meth_files_into_chunks(
  meth_files,
  meth_files_columns,
  file_grid_columns,
  meth_sites_df,
  collapse_strands,
  sequence_context,
  meth_site_groups,
  temp_chunk_dirs,
  zero_based,
  decimal_places,
  BPPARAM
)
}
\arguments{
\item{meth_files}{Paths to input methylation files.}

\item{meth_files_columns}{A list specifying the columns in meth_files.}

\item{file_grid_columns}{The grid column number for each file.}

\item{meth_sites_df}{A data.table with the positions of methylation sites.}

\item{collapse_strands}{TRUE or FALSE indicating whether or not to collapse data on + and - strands.}

\item{sequence_context}{A single character string or DNAString with the sequence context of the methylation sites e.g. CG or CHG.}

\item{meth_site_groups}{A list with the indices of the methylation sites in each group.}

\item{temp_chunk_dirs}{A vector giving the temporary directory associated with each chunk.}

\item{zero_based}{TRUE or FALSE indicating if files are zero-based.}

\item{decimal_places}{Integer indicating the number of decimal places to round beta values to.}

\item{BPPARAM}{A BiocParallelParam object.}
}
\value{
Invisibly returns NULL.
}
\description{
Split data from input methylation files into chunks
}
