% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_region_methylation.R
\name{.summarize_chunk_methylation}
\alias{.summarize_chunk_methylation}
\title{Summarize methylation values for regions in a chunk}
\usage{
.summarize_chunk_methylation(
  chunk_regions,
  meth_rse,
  assay,
  col_summary_function,
  na.rm,
  ...
)
}
\arguments{
\item{chunk_regions}{Chunk with genomic regions of interest.}

\item{meth_rse}{A RangedSummarizedExperiment with methylation values.}

\item{assay}{The assay from meth_rse to extract values from. Should be either an index or the name of an assay.}

\item{col_summary_function}{A function that summarizes column values.}

\item{na.rm}{TRUE or FALSE indicating whether to remove NA values when calculating summaries.}

\item{...}{Additional arguments to be passed to col_summary_function.}
}
\value{
A function which returns a list with the summarized methylation values for regions in each sample.
}
\description{
Summarize methylation values for regions in a chunk
}
