% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_meth_site_transcript_cors.R
\name{.tss_correlations}
\alias{.tss_correlations}
\title{Calculate meth site-transcript correlations for given TSS}
\usage{
.tss_correlations(correlation_objects)
}
\arguments{
\item{correlation_objects}{A list with a table of methylation values,
expression values for transcripts, a GRanges for the TSS and the name of the transcript.}
}
\value{
A data.frame with the correlation values
}
\description{
Calculate meth site-transcript correlations for given TSS
}
