% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_meth_site_transcript_cors.R
\name{.tss_iterator}
\alias{.tss_iterator}
\title{Create an iterator function for use with bpiterate}
\usage{
.tss_iterator(
  meth_values_chunk,
  tss_region_indices_list,
  transcript_values_list,
  tss_gr_chunk_list,
  cor_method,
  add_distance_to_region,
  min_number_complete_pairs,
  results_dir
)
}
\arguments{
\item{meth_values_chunk}{A table with methylation values for current chunk}

\item{tss_region_indices_list}{A list with the indices for methylation sites associated with each TSS.}

\item{transcript_values_list}{A list with expression values for transcripts.}

\item{tss_gr_chunk_list}{A list of GRanges with the TSS for the current chunk.}

\item{cor_method}{Correlation method to use.}

\item{add_distance_to_region}{TRUE or FALSE indicating whether to add distance to TSS.}

\item{min_number_complete_pairs}{The minimum number of complete pairs required to return a p-value for a correlation.}

\item{results_dir}{Location of results directory.}
}
\value{
An iterator function which returns a list with the parameters necessary for .tss_correlations.
}
\description{
Create an iterator function for use with bpiterate
}
