% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_meth_site_transcript_partial_cors.R
\name{.tss_iterator_partial_cors}
\alias{.tss_iterator_partial_cors}
\title{Create an iterator function for use with bpiterate}
\usage{
.tss_iterator_partial_cors(
  meth_values_chunk,
  meth_values_chunk_corrected,
  tss_region_indices_list,
  transcript_values,
  transcript_values_corrected,
  tss_for_chunk,
  cor_method,
  add_distance_to_region,
  results_dir,
  n_covariates
)
}
\arguments{
\item{meth_values_chunk}{A table with methylation values for current chunk}

\item{meth_values_chunk_corrected}{A table with corrected methylation values for current chunk}

\item{tss_region_indices_list}{A list with the indices for methylation sites associated with each TSS.}

\item{transcript_values}{A list with expression values for transcripts.}

\item{transcript_values_corrected}{A list with corrected expression values for transcripts.}

\item{tss_for_chunk}{A list of GRanges with the TSS for the current chunk.}

\item{cor_method}{Correlation method to use.}

\item{add_distance_to_region}{TRUE or FALSE indicating whether to add distance to TSS.}

\item{results_dir}{Location of results directory.}

\item{n_covariates}{Number of covariates to correct for}
}
\value{
An iterator function which returns a list with the parameters necessary for .tss_correlations.
}
\description{
Create an iterator function for use with bpiterate
}
