% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_datasets.R
\name{download_meth_dataset}
\alias{download_meth_dataset}
\title{Download a WGBS methylation dataset from TumourMethData}
\usage{
download_meth_dataset(dataset, dir = tempdir())
}
\arguments{
\item{dataset}{Name of the dataset to download. Must be one of the datsets listed in data(TumourMethDatasets).}

\item{dir}{Parent directory to create links to the HDF5 SummarizedExperiment dataset.
A subdirectory with the dataset name will be created within this directory. Default is tempdir().}
}
\value{
A RangedSummarizedExperiment with methylation values from the specified dataset.
}
\description{
Download a WGBS methylation dataset from TumourMethData
}
\examples{
mcrpc_wgbs_hg38_chr11 <- methodical::download_meth_dataset(dataset = "mcrpc_wgbs_hg38_chr11")
print(mcrpc_wgbs_hg38_chr11)
}
