% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_datasets.R
\name{downloadMethDataset}
\alias{downloadMethDataset}
\title{Download a WGBS methylation dataset from TumourMethData}
\usage{
downloadMethDataset(dataset, dir = tempdir())
}
\arguments{
\item{dataset}{Name of the dataset to download. Must be one of the datsets listed in data(TumourMethDatasets).}

\item{dir}{Parent directory to create links to the HDF5 SummarizedExperiment dataset.
A subdirectory with the dataset name will be created within this directory. Default is tempdir().
If a directory named \code{paste(dir, dataset, sep = "/")} already exists,
will attempt to load a HDF5 summarized experiment from there and return an error if this fails.}
}
\value{
A RangedSummarizedExperiment with methylation values from the specified dataset.
}
\description{
Download a WGBS methylation dataset from TumourMethData
}
\examples{
mcrpc_wgbs_hg38_chr11 <- methodical::downloadMethDataset(dataset = "mcrpc_wgbs_hg38_chr11")
print(mcrpc_wgbs_hg38_chr11)
}
