% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmr_functions.R
\docType{data}
\name{example_meth_transcript_cors}
\alias{example_meth_transcript_cors}
\title{example_meth_transcript_cors}
\format{
A data.frame with 5 columns giving the name of the CpG site (meth_site), name of the transcript associated with the TSS,
Spearman correlation value between the methylation of the CpG site and expression of the transcript,
p-value associated with the correlations and distance from the CpG site to the TSS.
}
\usage{
example_meth_transcript_cors
}
\description{
A data.frame with the correlation results for CpG sites within +/- 5 KB of the TUBB6 (ENST00000591909) TSS.
}
\keyword{datasets}
