% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meth_rse_functions.R
\name{export_bedGraph_from_rse}
\alias{export_bedGraph_from_rse}
\title{Export values for a sample in a RangedSummarizedExperiment as a bedGraph}
\usage{
export_bedGraph_from_rse(meth_rse, assay_number = 1, sample_name, file_name)
}
\arguments{
\item{meth_rse}{A RangedSummarizedExperiment for methylation data.}

\item{assay_number}{The assay from meth_rse to extract values from. Default is the first assay.}

\item{sample_name}{The name of a single sample in meth_rse.}

\item{file_name}{The output filename.}
}
\value{
A data.frame with the methylation site values for all sites in meth_rse which overlap genomic_ranges.
Row names are the coordinates of the sites as a character vector.
}
\description{
Export values for a sample in a RangedSummarizedExperiment as a bedGraph
}
