% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meth_rse_functions.R
\name{export_files_from_rse}
\alias{export_files_from_rse}
\title{Export values for samples in a RangedSummarizedExperiment as bedGraphs or bigWigs}
\usage{
export_files_from_rse(
  meth_rse,
  assay_number = 1,
  samples = NULL,
  filetype,
  output_dir = ".",
  compress_bedGraphs = TRUE
)
}
\arguments{
\item{meth_rse}{A RangedSummarizedExperiment for methylation data.}

\item{assay_number}{The assay from meth_rse to extract values from. Default is the first assay.}

\item{samples}{The names of one or more samples in meth_rse. Default is all samples.}

\item{filetype}{The type of file to output. Should be one of either "bedGraph" or "bigWig".
If meth_rse contains a large number of regions, may be more efficient to export as bedGraph and convert to bigWigs externally.}

\item{output_dir}{The directory to save the output files. Default is current directory.}

\item{compress_bedGraphs}{Whether or not to compress output bedGraph files with gzip. Default is TRUE.}
}
\description{
Export values for samples in a RangedSummarizedExperiment as bedGraphs or bigWigs
}
