% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{extract_cpg_values_from_methrix}
\alias{extract_cpg_values_from_methrix}
\title{Get methylation values of CpG sites overlapping a GRanges object from a methrix object}
\usage{
extract_cpg_values_from_methrix(
  methrix,
  genomic_regions = NULL,
  samples = NULL
)
}
\arguments{
\item{methrix}{A methrix object}

\item{genomic_regions}{Genomic regions to subset by. Could be a data.table with 3 columns (chr, start, end) or a GenomicRanges object. Default is to extract all CpG sites in the methrix.}

\item{samples}{Sample names to subset by. Default is to get values for all samples.}
}
\value{
A data.frame with the methylation values of overlapping CpG sites with CpG names as first column
}
\description{
Get methylation values of CpG sites overlapping a GRanges object from a methrix object
}
