% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meth_rse_functions.R
\name{extract_granges_meth_site_values}
\alias{extract_granges_meth_site_values}
\title{Extract values for methylation sites overlapping genomic regions from a methylation RSE.}
\usage{
extract_granges_meth_site_values(
  meth_rse,
  genomic_regions,
  samples_subset = NULL,
  assay_number = 1
)
}
\arguments{
\item{meth_rse}{A RangedSummarizedExperiment for methylation data.}

\item{samples_subset}{Optional sample names used to subset meth_rse.}

\item{assay_number}{The assay from meth_rse to extract values from. Default is the first assay.}

\item{genomic_ranges}{A GRanges object.}
}
\value{
A data.frame with the methylation site values for all sites in meth_rse which overlap genomic_ranges.
Row names are the coordinates of the sites as a character vector.
}
\description{
Extract values for methylation sites overlapping genomic regions from a methylation RSE.
}
