% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{find_cpgs_overlapping_granges}
\alias{find_cpgs_overlapping_granges}
\title{Get names of CpGs in a methrix object which overlap a GRanges object}
\usage{
find_cpgs_overlapping_granges(methrix_object, granges_object)
}
\arguments{
\item{methrix_object}{A methrix object}

\item{genomic_regions}{A GRanges object}
}
\description{
Get names of CpGs in a methrix object which overlap a GRanges object
}
