% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_annotation.R
\docType{data}
\name{genome_annotation_hg38}
\alias{genome_annotation_hg38}
\title{genome_annotation_hg38}
\format{
GRanges object with 9,034,338 ranges and one metadata column region_type describing the region.
}
\source{
Created by combining gene annotation from Gencode v38, regulatory feature annotation from Ensembl release 109 and
masked CpG island data and repeat annotation from UCSC genome browser annotation database.
}
\usage{
genome_annotation_hg38
}
\description{
Comprehensive annotation of the standard chromosomes of the hg38 genome build including
gene annotation, structural annotation, regulatory element annotation and repeat annotation.
}
\keyword{datasets}
