% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_documentation.R
\docType{data}
\name{hg19_chr18_cpgs}
\alias{hg19_chr18_cpgs}
\title{hg19_chr18_cpgs}
\format{
A GRanges object.
}
\usage{
hg19_chr18_cpgs
}
\description{
A GRanges object with CpG sites on chr18 for hg19
}
\keyword{datasets}
