% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_documentation.R
\docType{data}
\name{hg38_chr18}
\alias{hg38_chr18}
\title{hg38_chr18}
\format{
A GRanges object.
}
\usage{
hg38_chr18
}
\description{
A DNAStringSet for first 1,000,000 bp from chr18 from BSgenome.Hsapiens.UCSC.hg38
}
\keyword{datasets}
