% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_documentation.R
\docType{data}
\name{hg38_cpg_islands}
\alias{hg38_cpg_islands}
\title{hg38_cpg_islands}
\format{
A GRanges object.
}
\usage{
hg38_cpg_islands
}
\description{
A GRanges with CpG islands, shelves and shores for hg38
}
\keyword{datasets}
