% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_documentation.R
\docType{data}
\name{hg38_cpgs_subset}
\alias{hg38_cpgs_subset}
\title{hg38_cpgs_subset}
\format{
A GRanges object.
}
\usage{
hg38_cpgs_subset
}
\description{
All the CpG sites within the first one million base pairs of chromosome 1.
}
\keyword{datasets}
