% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_meth_rse.R
\docType{data}
\name{infinium_450k_probe_granges_hg38}
\alias{infinium_450k_probe_granges_hg38}
\title{infinium_450k_probe_granges_hg38}
\format{
GRanges object with 480,975 ranges and one metadata column name giving the name of the associated probe.
}
\usage{
infinium_450k_probe_granges_hg38
}
\description{
The genomic coordinates for infinium_450k_probe_granges_hg19 lifted-over to hg38.
Probes which could not be mapped to hg38 and those resulting in one-to-many or many-to-one mappings
were removed as well as those which did not overlap CpG sites after lift-over.
}
\keyword{datasets}
