% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{liftover_cpgs}
\alias{liftover_cpgs}
\title{Liftover CpGs sites from one genome build to another.}
\usage{
liftover_cpgs(cpg_values, liftover_chain)
}
\arguments{
\item{cpg_values}{A data.frame with methylation values for CpG values in different samples such as returned by extract_cpg_values_from_methrix().}

\item{liftover_chain}{A liftover Chain object specific to the genome build of the CpG sites in cpg_values and the desired target genome build.}
}
\value{
A data.frame with CpG methylation values for those CpGs which could be uniquely lifted over to the genome build indicated by liftover_chain.
CpG sites which map to multiple sites in the target genome build or which are mapped to the same site as another CpG site are excluded.
}
\description{
Liftover CpGs sites from one genome build to another.
}
