% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meth_rse_functions.R
\name{liftover_meth_rse}
\alias{liftover_meth_rse}
\title{Liftover rowRanges of a RangedSummarizedExperiment for methylation data from one genome build to another}
\usage{
liftover_meth_rse(
  meth_rse,
  chain,
  remove_one_to_many_mapping = TRUE,
  permitted_target_regions = NULL
)
}
\arguments{
\item{meth_rse}{A RangedSummarizedExperiment for methylation data}

\item{chain}{A "Chain" object to be used with rtracklayer::liftOver}

\item{remove_one_to_many_mapping}{A logical value indicating whether to remove regions in the source genome
which map to multiple regions in the target genome. Default is TRUE.}

\item{permitted_target_regions}{An optional GRanges object used to filter the rowRanges by overlaps after liftover,
for example CpG sites from the target genome. Any regions which do not overlap permitted_target_regions will be removed.
GRangesList to GRanges if all remaining source regions can be uniquely mapped to the target genome.}
}
\value{
A RangedSummarizedExperiment with rowRanges lifted over to the genome build indicated by chain.
}
\description{
Removes methylation sites which cannot be mapped to the target genome build and those which result in
many-to-one mappings. Also removes one-to-many mappings by default and can remove sites which do not
map to allowed regions in the target genome e.g. CpG sites.
}
