% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_meth_rse.R
\name{make_meth_rse_from_array_files}
\alias{make_meth_rse_from_array_files}
\title{Create a HDF5-backed RangedSummarizedExperiment for methylation values in array files}
\usage{
make_meth_rse_from_array_files(
  array_files,
  probe_name_column = 1,
  beta_value_column = 2,
  convert_percentages = T,
  decimal_places = NA,
  probe_ranges,
  sample_metadata = NULL,
  hdf5_dir,
  dataset_name = "beta",
  overwrite = F,
  chunkdim = NULL,
  temporary_dir,
  ncores = 1,
  ...
)
}
\arguments{
\item{array_files}{A vector of paths to bedGraph files. Automatically detects if array_files contain a header if every field in the first line is a character.}

\item{probe_name_column}{The number of the column which corresponds to the name of the probes. Default is 1st column.}

\item{beta_value_column}{The number of the column which corresponds to the beta values . Default is 2nd column.}

\item{convert_percentages}{A logical value indicating whether bedGraph values should be converted from percentages to proportions if the maximum value is greater than 1.
Default is TRUE.}

\item{decimal_places}{Integer indicating the number of decimal places to round beta values to. Default is 2.}

\item{probe_ranges}{A GRanges object giving the genomic locations of probes where each region corresponds to a separate probe.
There should be a metadata column called name with the name of the probe associated with each region.
Any probes in array_files that are not in probe_ranges are ignored.}

\item{sample_metadata}{Sample metadata to be used as colData for the RangedSummarizedExperiment}

\item{hdf5_dir}{Directory to save HDF5 file. Is created if it doesn't exist. HDF5 file is called assays.h5.}

\item{dataset_name}{Name to give data set in HDF5 file. Default is "beta".}

\item{overwrite}{Logical value indicating whether to allow overwriting if dataset_name already exists in assays.h5. Default is FALSE.}

\item{chunkdim}{The dimensions of the chunks for the HDF5 file. Should be a vector of length 2 with giving the number of rows and then the number of columns in each chunk.}

\item{temporary_dir}{Name to give a temporary directory to store intermediate files. A directory with this name cannot already exist.
Will be deleted after completion.}

\item{ncores}{Number of array files to process at a time in parallel. Default is 1.}

\item{...}{Additional arguments to be passed to HDF5Array::HDF5RealizationSink() for controlling the physical properties of the created HDF5 file,
such as compression level. Uses the defaults for any properties that are not specified.}
}
\value{
A RangedSummarizedExperiment with methylation values for all methylation sites in meth_sites. Methylation sites will be in the same order as sort(meth_sites).
}
\description{
Create a HDF5-backed RangedSummarizedExperiment for methylation values in array files
}
