% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_methrix_from_array_files.R
\name{make_methrix_from_array_files}
\alias{make_methrix_from_array_files}
\title{Create a Methrix object from Illumina methylation array files.}
\usage{
make_methrix_from_array_files(
  array_files,
  probe_name_column = 1,
  beta_value_column = 2,
  methylation_array,
  genome,
  coldata = NULL,
  h5_dir,
  tmpdir = tempdir()
)
}
\arguments{
\item{array_files}{A list of paths to Illumina methylation array files}

\item{probe_name_column}{The number of the column which corresponds to the name of the probes. Default is 1st column.}

\item{beta_value_column}{The number of the column which corresponds to the beta values . Default is 2nd column.}

\item{methylation_array}{Name of the Illumina methylation array. At the moment can just be "HumanMethylation450K".}

\item{genome}{Name of a BSgenome. Allowed values are "BSgenome.Hsapiens.UCSC.hg19", "BSgenome.Hsapiens.UCSC.hg38" and "BSgenome.Hsapiens.NCBI.GRCh38".}

\item{coldata}{A data.frame with sample metadata. row.names should be the names of the samples.}

\item{h5_dir}{Directory to store the HDF5 methrix object. Cannot already exist.}

\item{tmpdir}{Directory in which to create a subdirectory to store temporary files. Defaults to tempdir().}
}
\value{
The methrix object created from array_files
}
\description{
Needs to use 1.4 times the space used by array_files as temporary space
}
