% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_shiny_app.R
\name{make_shiny_app}
\alias{make_shiny_app}
\title{Created a Shiny app for displaying CpG methylation and transcript expression correlations}
\usage{
make_shiny_app(
  methrix_object,
  transcript_table,
  correlation_table,
  sample_groups = NULL
)
}
\arguments{
\item{methrix_object}{A methrix object}

\item{transcript_table}{A data.frame with transcript expression values where row names are the names of the transcripts and column names are sample names.}

\item{correlation_table}{A table with the calculated correlations for methylation of CpG sites and transcript expression,
such as returned by combine_transcription_cpg_correlations()}

\item{sample_groups}{An optional factor giving the names of the groups that samples belong to.}
}
\value{
A Shiny App which creates scatter plots for input CpG names and transcripts or else from selected rows in the table.
}
\description{
Created a Shiny app for displaying CpG methylation and transcript expression correlations
}
