% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_meth_rse.R
\name{makeMethRSEFrominput_files}
\alias{makeMethRSEFrominput_files}
\title{Create a HDF5-backed RangedSummarizedExperiment for methylation values in input_files}
\usage{
makeMethRSEFrominput_files(
  input_files,
  seqnames_column = 1,
  start_column = 2,
  end_column = 3,
  methylation_column = 5,
  methylation_column_is_counts = "counts",
  coverage_column = 6,
  zero_based = TRUE,
  normalization_factor = NULL,
  decimal_places = NA,
  meth_sites,
  sample_metadata = NULL,
  hdf5_dir,
  dataset_name = "beta",
  overwrite = FALSE,
  chunkdim = NULL,
  temporary_dir = NULL,
  BPPARAM = BiocParallel::SerialParam(),
  ...
)
}
\arguments{
\item{input_files}{A vector of paths to input files. Automatically detects if input_files contain a header if every field in the first line is a character.}

\item{seqnames_column}{The column number in input_files which corresponds to the sequence names. Default is 1st column.}

\item{start_column}{The column number in input_files which corresponds to the start positions. Default is 2nd column.}

\item{end_column}{The column number in input_files which corresponds to the end positions. Default is 3rd column.}

\item{methylation_column}{The column number in input_files which corresponds to the methylation values. Default is 5th column.}

\item{methylation_column_is_counts}{A logical value indicating of methylation_column gives the number of reads supporting methylation.
If FALSE, methylation_column is assumed to be the proportion of methylated reads. Default is TRUE.}

\item{coverage_column}{The column number in input_files which corresponds to the total coverage. Default is the 6th column.}

\item{zero_based}{TRUE or FALSE indicating if files are zero-based. Default value is TRUE.}

\item{normalization_factor}{An optional numerical value to divide methylation values by to convert them to fractions e.g. 100 if they are percentages.
Default is to leave values as they are in the input files.}

\item{decimal_places}{Optional integer indicating the number of decimal places to round beta values to. Default is not to round.}

\item{meth_sites}{A GRanges object with the locations of the methylation sites of interest. Any methylation sites in input_files that are not in meth_sites are ignored.}

\item{sample_metadata}{Sample metadata to be used as colData for the RangedSummarizedExperiment.}

\item{hdf5_dir}{Directory to save HDF5 file. Is created if it doesn't exist. HDF5 file is called assays.h5.}

\item{dataset_name}{Name to give data set in HDF5 file. Default is "beta".}

\item{overwrite}{TRUE or FALSE indicating whether to allow overwriting if dataset_name already exists in assays.h5. Default is FALSE.}

\item{chunkdim}{The dimensions of the chunks for the HDF5 file. Should be a vector of length 2 giving the number of rows and then the number of columns in each chunk.
Uses HDF5Array::getHDF5DumpChunkDim(length(meth_sites), length(input_files))) by default.}

\item{temporary_dir}{Name to give temporary directory created to store intermediate files. A directory with this name cannot already exist.
Default is to create a name using tempfile("temporary_meth_chunks_").
Will be deleted after completion.}

\item{BPPARAM}{A BiocParallelParam object for parallel processing. Defaults to \code{BiocParallel::SerialParam()}.}

\item{...}{Additional arguments to be passed to HDF5Array::HDF5RealizationSink() for controlling the physical properties of the created HDF5 file,
such as compression level. Uses the defaults for any properties that are not specified.}
}
\value{
A RangedSummarizedExperiment with methylation values for all methylation sites in meth_sites. methylation sites will be in the same order as sort(meth_sites).
}
\description{
Create a HDF5-backed RangedSummarizedExperiment for methylation values in input_files
}
\examples{

# Load CpGs within first million base pairs of chromosome 1 as a GRanges object
data("hg38_cpgs_subset", package = "methodical")

# Get paths to input_files
input_files <- list.files(path = system.file('extdata', package = 'methodical'), 
  pattern = ".bg.gz", full.names = TRUE)

# Create sample metadata
sample_metadata <- data.frame(
  tcga_project = gsub("_.*", "", gsub("TCGA_", "", basename(input_files))),
  sample_type = ifelse(grepl("N", basename(input_files)), "Normal", "Tumour"),
  row.names = tools::file_path_sans_ext(basename(input_files))
)

# Create a HDF5-backed RangedSummarizedExperiment from input_files
meth_rse <- makeMethRSEFrominput_files(input_files = input_files, 
  meth_sites = hg38_cpgs_subset, sample_metadata = sample_metadata, 
  hdf5_dir = paste0(tempdir(), "/bedgraph_hdf5_1"))
  
}
