% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meth_rse_functions.R
\name{mask_ranges_in_rse}
\alias{mask_ranges_in_rse}
\title{Mask regions in a ranged summarized experiment}
\usage{
mask_ranges_in_rse(rse, mask_ranges, assay = 1)
}
\arguments{
\item{rse}{A RangedSummarizedExperiment.}

\item{mask_ranges}{Either a GRanges with regions to be masked in all samples or a GRangesList object with
different regions to mask in each sample. If using a GRangesList object, names of the list
elements should be the names of samples in rse.}

\item{assay}{Assay to perform masking. Default is first assay}
}
\value{
A RangedSummarizedExperiment with the regions present in mask_ranges masked
}
\description{
Mask regions in a ranged summarized experiment
}
