% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{methylation_feature_scatter_plot}
\alias{methylation_feature_scatter_plot}
\title{Create a scatter plot of methylation and another feature}
\usage{
methylation_feature_scatter_plot(
  cpg_values,
  cpg_name,
  feature,
  sample_groups = NULL,
  group_colours = NULL,
  use_names = T,
  regression_line = F,
  add_correlation = F,
  method = "p",
  position = NULL,
  label_size = 3,
  xlab = "Methylation Value",
  ylab = "Feature Value",
  title = "Methylation Correlation",
  alpha = 1
)
}
\arguments{
\item{cpg_values}{A data.frame with the methylation values of CpGs for different samples such as returned by extract_cpg_values_from_methrix.}

\item{cpg_name}{The name of a CpG site in cpg_values.}

\item{feature}{A named numeric vector for a feature of interest.}

\item{sample_groups}{An optional factor giving the names of the groups that samples belong to.}

\item{use_names}{A logical value indicating whether elements in cpg_values and feature should be matched by their names. Default is TRUE.}

\item{regression_line}{A logical value indicating whether to add a regression line to the plot. Default is FALSE.}

\item{add_correlation}{A logical value indicating whether to add correlation to the plot. Default if FALSE.}

\item{method}{A character string indicating which correlation coefficient is to be computed. Identical to methods from cor().}

\item{position}{A numeric vector of length 2 giving the x and y coordinates on a 0 to 1 scale of the correlation label.
Default is 90\% of max methylation value for the x-coordinate and 0.99 for the y coordinate.}

\item{label_size}{Size of the add_correlation. Default is 3.}

\item{xlab}{x-axis label. Default is "Methylation Value".}

\item{ylab}{y-axis label. Default is}

\item{title}{Title of the plot}

\item{alpha}{Value specifying the alpha}

\item{group_colors}{An optional vector of colours to be used for each group. Default is to use default ggplot2 colours.}
}
\value{
A ggplot object
}
\description{
Calculates the specified correlation coefficient of two named numeric vectors, matching elements by common names and ignoring elements whose name is not present in both vectors.
}
