% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_cpg_methylation}
\alias{plot_cpg_methylation}
\title{Create plot of methylation values}
\usage{
plot_cpg_methylation(
  cpg_values,
  geom,
  samples_subset = NULL,
  sample_groups = NULL,
  group_summary_function = NULL,
  group_difference = NULL,
  group_colours = NULL,
  reference_region = NULL,
  title = NULL,
  xlabel = "Genomic Position",
  ylabel = "CpG Methylation"
)
}
\arguments{
\item{cpg_values}{A data.frame with the methylation values of CpGs for different samples such as returned by extract_cpg_values_from_methrix.
All CpG sites must be located on the same sequence.}

\item{geom}{geom to display. Can be one of geom_point, geom_line, geom_boxplot or geom_linepoint which indicates to use both geom_point and geom_line together.}

\item{samples_subset}{An optional vector of sample names to use to subset cpg_values.}

\item{sample_groups}{An optional factor vector giving the names of the groups that samples belong to. Alternatively, can set to "all_samples" to plot and colour each sample separately.}

\item{group_summary_function}{The name of a function to use to summarize values from each group before plotting. Can be any function which returns a single value from a numeric vector e.g. mean.
If not set, individual samples are plotted instead of group summaries.}

\item{group_difference}{An optional vector of length two with the names of two groups from sample_groups. Providing group_difference specifies to plot the difference between
the summary values calculated for each group using group_summary_function by subtracting the second group from the first.}

\item{reference_region}{An optional GRanges object with a single range. If provided, the x-axis will show the distance of CpG sites to the start of this region with CpGs upstream
relative to the reference_region shown first. If not, the x-axis will show the start site coordinate of the CpG.}

\item{title}{Title of the plot. Default is no title.}

\item{xlabel}{Label for the X axis in the plot. Default is "Genomic Position".}

\item{ylabel}{Label for the Y axis in the plot. Default is "CpG Methylation".}

\item{group_colors}{An optional vector of colours to be used for each group. Default is to use default ggplot2 colours. If group_difference is provided, should just be a single colour.}
}
\value{
A ggplot object of the mean methylation values of the CpGs in cpg_values.
}
\description{
Create plot of methylation values
}
