% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_methodical_scores}
\alias{plot_methodical_scores}
\title{Create plot of Methodical score values for methylation sites around a TSS}
\usage{
plot_methodical_scores(
  meth_site_values,
  reference_region = NULL,
  title = NULL,
  xlabel = "Genomic Position",
  low_colour = "#7B5C90",
  high_colour = "#BFAB25",
  p_value_threshold = 0.005
)
}
\arguments{
\item{meth_site_values}{A data.frame with correlation values for methylation sites. There should be one column called "cor".
and another called "p_val" which are used to calculate the Methodical score. row.names should be the names of methylation sites and all methylation sites must be located on the same sequence.}

\item{reference_region}{An optional GRanges object with a single range. If provided, the x-axis will show the distance of methylation sites to the start of this region with methylation sites upstream.
relative to the reference_region shown first. If not, the x-axis will show the start site coordinate of the methylation site.}

\item{title}{Title of the plot. Default is no title.}

\item{xlabel}{Label for the X axis in the plot. Default is "Genomic Position".}

\item{low_colour}{Colour to use for low values. Default value is "#7B5C90".}

\item{high_colour}{Colour to use for high values. Default value is "#BFAB25".}

\item{p_value_threshold}{The p-value threshold used to identify TMRs. Default value is 0.005. Set to NULL to not display thresholds.}
}
\value{
A ggplot object
}
\description{
Create plot of Methodical score values for methylation sites around a TSS
}
