% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_smoothed_methodical_scores}
\alias{plot_smoothed_methodical_scores}
\title{Add smoothed methodical scores curve to a methylation site plot}
\usage{
plot_smoothed_methodical_scores(
  meth_site_plot,
  smoothed_methodical_scores,
  colour = "black",
  linewidth = 1,
  alpha = 0.75
)
}
\arguments{
\item{meth_site_plot}{A plot of methylation site values around a TSS.}

\item{smoothed_methodical_scores}{A vector with smoothed methodical scores for each methylation site in meth_site_plot.}

\item{colour}{Colour of the smoothed curve. Default is "black"}

\item{linewidth}{Line width of the smoothed curve. Default value is 1.}

\item{alpha}{Alpha value for the curve. Default value is 0.75.}
}
\value{
A ggplot object
}
\description{
Add smoothed methodical scores curve to a methylation site plot
}
