% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plotMethSiteValues}
\alias{plotMethSiteValues}
\title{Create plot of methylation values}
\usage{
plotMethSiteValues(
  meth_site_values,
  column_name,
  reference_tss = FALSE,
  title = NULL,
  xlabel = NULL,
  ylabel = "Value",
  value_colours = "set1"
)
}
\arguments{
\item{meth_site_values}{A data.frame with values associated with methylation sites.
First column should be called meth_site and give the names of methylation sites.
All methylation sites must be located on the same sequence.}

\item{column_name}{Name of column in meth_site_values to plot.}

\item{reference_tss}{A logical value indicating whether to show distances on the X-axis
relative to the TSS stored as an attribute \code{tss_range} of meth_site_values.
Alternatively, can provide a GRanges object with a single range for such a TSS site.
In either case, will show the distance of methylation sites to the start of this region with methylation sites upstream
relative to the reference_tss shown first.
If FALSE (the default), the x-axis will instead show the start site coordinate of the methylation site.}

\item{title}{Title of the plot. Default is no title.}

\item{xlabel}{Label for the X axis in the plot. Defaults to "Distance to TSS" if reference_tss is used or
"seqname position" where seqname is the name of the relevant sequence.}

\item{ylabel}{Label for the Y axis in the plot. Default is "Value".}

\item{value_colours}{A vector with two colours to use, one for low values and the other for high values.
Alternatively, can use one of two predefined colour sets by providing either "set1" or "set2":
set1 uses "#53868B" (blue) for low values and "#CD2626" (red) for high values
while set2 uses "#7B5C90" (purple) for low values and ""#bfab25" (gold) for high values. Default is "set1".}
}
\value{
A ggplot object
}
\description{
Create plot of methylation values
}
\examples{
# Load methylation-transcript correlation results for TUBB6 gene
data("tubb6_cpg_meth_transcript_cors", package = "methodical")

# Plot methylation-transcript correlation values around TUBB6 TSS
methodical::plotMethSiteValues(tubb6_cpg_meth_transcript_cors, column_name = "cor", ylabel = "Spearman Correlation")

# Create same plot but showing the distance to the TUBB6 TSS on the x-axis
methodical::plotMethSiteValues(tubb6_cpg_meth_transcript_cors, column_name = "cor", 
  ylabel = "Spearman Correlation", reference_tss = attributes(tubb6_cpg_meth_transcript_cors)$tss_range)

}
