% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/granges_manipulation.R
\name{relative_ranges}
\alias{relative_ranges}
\title{Find locations of ranges relative to a reference positions}
\usage{
relative_ranges(genomic_regions, reference_positions)
}
\arguments{
\item{genomic_regions}{A GRanges object}

\item{reference_positions}{A GRanges object. Each range should have width 1. Upstream and downstream are relative to reference_positions}
}
\value{
A GRanges object
}
\description{
Find locations of ranges relative to a reference positions
}
