% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{sample_cpgs}
\alias{sample_cpgs}
\title{Sample random CpGs from a methrix object}
\usage{
sample_cpgs(
  methrix_object,
  n_cpgs = 1000,
  regions = NULL,
  contigs = NULL,
  samples = NULL,
  overlap_type = "any"
)
}
\arguments{
\item{methrix_object}{A methrix object}

\item{n_cpgs}{Number of CpG sites to randomly select. Default is 1000.}

\item{regions}{Genomic regions to subset by. Could be a data.table with 3 columns (chr, start, end) or a GenomicRanges object}

\item{contigs}{Chromosome names to subset by.}

\item{samples}{Sample names to subset by.}

\item{overlap_type}{defines the type of the overlap of the CpG sites with the target region.
Default value is 'any'. For detailed description, see the foverlaps function of the data.table package.}
}
\description{
Sample random CpGs from a methrix object
}
