% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_cluster.R
\name{setup_cluster}
\alias{setup_cluster}
\title{Setup a cluster for use with doParallel}
\usage{
setup_cluster(ncores, packages = NULL, outfile = NULL)
}
\arguments{
\item{ncores}{Number of cores.}

\item{packages}{A character vector with the names of packages to load.}

\item{outfile}{Outfile for the cluster. Default is the NULL device.}
}
\value{
A SOCKcluster if ncores is greater than 1 or NULL otherwise.
}
\description{
Setup a cluster for use with doParallel
}
