% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_meth_rse_accessory_functions.R
\name{split_meth_array_files_into_chunks}
\alias{split_meth_array_files_into_chunks}
\title{Split data from methylation array files into chunks}
\usage{
split_meth_array_files_into_chunks(
  array_files,
  probe_name_column,
  beta_value_column,
  file_grid_columns,
  probe_ranges,
  probe_groups,
  temp_chunk_dirs,
  convert_percentages,
  decimal_places,
  ncores
)
}
\arguments{
\item{array_files}{Paths to methylation array files.}

\item{probe_name_column}{The column number in array_files which corresponds to the name of the probes. Default is 1st column.}

\item{beta_value_column}{The column number in array_files which corresponds to the beta values. Default is 2nd column.}

\item{file_grid_columns}{The grid column number for each file.}

\item{probe_ranges}{A GRanges object giving the genomic locations of probes where each region corresponds to a separate probe.}

\item{probe_groups}{A list with the indices of the probes in each group.}

\item{temp_chunk_dirs}{A vector giving the temporary directory associated with each chunk.}

\item{convert_percentages}{A logical value indicating whether bedGraph values should be converted from percentages to proportions if the maximum value is greater than 1.}

\item{decimal_places}{Integer indicating the number of decimal places to round beta values to.}

\item{ncores}{Number of cores to use.}
}
\value{
A data.table with the probe sites sorted by seqnames, start and probe name.
}
\description{
Split data from methylation array files into chunks
}
