% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/granges_manipulation.R
\name{stranded_distance}
\alias{stranded_distance}
\title{Calculate distances of query GRanges upstream or downstream of subject GRanges}
\usage{
stranded_distance(query_gr, subject_gr)
}
\arguments{
\item{query_gr}{A GRanges object}

\item{subject_gr}{A GRanges object.}
}
\value{
A numeric vector of distances
}
\description{
Upstream and downstream are relative to the strand of subject_gr. Unstranded regions are treated the same as regions on the "+" strand.
}
