% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kallisto_quantify.R
\name{sum_transcript_values_for_genes}
\alias{sum_transcript_values_for_genes}
\title{Combine the expression values of transcripts to get overall expression of their associated genes}
\usage{
sum_transcript_values_for_genes(
  transcript_expression_table,
  gene_to_transcript_list
)
}
\arguments{
\item{transcript_expression_table}{A table where rows are transcripts and columns are samples. Row names should be the names of transcripts.}

\item{gene_to_transcript_list}{A list of vectors where the name of each list entry is a gene name and its elements are the names of transcripts.
Can alternatively be a GRangeList where the name of each list element is a gene and the names of the individual ranges are transcripts.}
}
\value{
A data.frame with the sum of transcript expression values for genes where rows are genes and columns are samples
}
\description{
Combine the expression values of transcripts to get overall expression of their associated genes
}
