% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{summarize_region_methylation}
\alias{summarize_region_methylation}
\title{Summarize methylation of CpGs overalpping regions}
\usage{
summarize_region_methylation(
  methrix_object,
  genomic_regions,
  genomic_regions_names = NULL,
  summary_function = "mean",
  overlap_type = "any"
)
}
\arguments{
\item{methrix_object}{A methrix object}

\item{genomic_regions}{Genomic regions to calculate mean CpG methylation for.}

\item{genomic_regions_names}{Names to give genomic_regions. Default is to name them region_1, region_2, etc. if no names provided.}

\item{summary_function}{How methylation in regions should be summarized. Can be one of "mean", "sum", "max" or "min". Default is "mean".}

\item{overlap_type}{Type of overlap_type to use with CpG region overlaps. Same as the options for methrix::get_region_summary(), but default is "any".}
}
\value{
A data.frame with the mean methylation of CpGs in each region for each sample
}
\description{
Summarize methylation of CpGs overalpping regions
}
