% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcript_cpg_correlations2.R
\name{transcription_cpg_correlations}
\alias{transcription_cpg_correlations}
\title{Calculate correlation between methylation of CpGs surrounding genomic regions and transcription of associated features}
\usage{
transcription_cpg_correlations(
  methrix_object,
  transcript_expression_table,
  subset_samples = NULL,
  genomic_regions,
  expand_upstream = 0,
  expand_downstream = 0,
  cor_method = "pearson",
  n_permutations = 10000,
  add_distance_to_region = T,
  ncores = 1,
  random_seed = 123
)
}
\arguments{
\item{methrix_object}{A methrix object. Names of samples must match those in transcript_expression_table unless subset_samples provided.}

\item{transcript_expression_table}{A table with the expression values for transcripts, where rows correspond to transcripts and columns to samples.
There should be one row for each region in genomic_regions. Names of samples must match those in methrix_object unless subset_samples provided.}

\item{subset_samples}{Sample names to subset methrix object and transcript_expression_table. Provided samples must be found in both methrix_object and transcript_expression_table.
Default is to use all samples in methrix_object and transcript_expression_table.}

\item{genomic_regions}{Genomic regions to calculate mean CpG methylation for. There should be one region for each row in transcript_expression_table.
Must have a metadata column called transcript_id that is identical to the row.names of transcript_expression_table.}

\item{expand_upstream}{Number of bases to add upstream of each region in genomic regions. Must be numeric vector of length 1 or equal to the length of genomic_regions. Default is 0.}

\item{expand_downstream}{Number of bases to add downstream of each region in genomic regions. Must be numeric vector of length 1 or equal to the length of genomic_regions. Default is 0.}

\item{cor_method}{A character string indicating which correlation coefficient is to be computed. Identical to methods from cor(). Default is "pearson".}

\item{n_permutations}{Number of permutations for permuation test. Default is 10,000.}

\item{add_distance_to_region}{A logical value indicating whether to add the distance of CpG sites to the center of the associated genomic region. Default value is TRUE.
Setting to FALSE will roughly half the total running time.}

\item{ncores}{Number of cores to use. Default is 1.}

\item{random_seed}{Integer to use as a random seed. Default is 123.}
}
\value{
A list of data.frames with the correlation of CpG sites surrounding a specified genomic region with a given feature, p-values and adjusted q-values for the correlations.
Distance of the CpG upstream or downstream to the center of the region is also provided.
}
\description{
AT THE MOMENT, IT RELIES ON transcript_expression_table AND genomic_regions HAVING MATCHING TRANSCRIPT IDS. MAYBE CHANGE THIS.
MAYBE ADD OPTION NOT TO ADD DISTANCE OR SIGNIFICANCE
ADD PARAMETER WITH DESIDRED CHROMOSOME ORDER MAYBE
ADD FUNCTION TO COMBINE RESULTS INTO A SINGLE TABLE
Sort by distance? Maybe not if it increases running time a lot
index using numbers rather than characters
}
