% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_documentation.R
\docType{data}
\name{tubb6_correlation_plot}
\alias{tubb6_correlation_plot}
\title{tubb6_correlation_plot}
\format{
A ggplot object.
}
\usage{
tubb6_correlation_plot
}
\description{
A plot of the correlation values between methylation-transcription correlations for CpG
sites around the TUBB6 TSS.
}
\keyword{datasets}
