% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_documentation.R, R/tmr_functions.R
\docType{data}
\name{tubb6_cpg_meth_transcript_cors}
\alias{tubb6_cpg_meth_transcript_cors}
\title{tubb6_cpg_meth_transcript_cors}
\format{
A ggplot object.

A data.frame with 5 columns giving the name of the CpG site (meth_site), name of the transcript associated with the TSS,
Spearman correlation value between the methylation of the CpG site and expression of the transcript,
p-value associated with the correlations and distance from the CpG site to the TSS.
}
\usage{
tubb6_cpg_meth_transcript_cors

tubb6_cpg_meth_transcript_cors
}
\description{
A data.frame with the methylation-transcription correlation results for CpGs arond the TUBB6 TSS.

A data.frame with the correlation results for CpG sites within +/- 5 KB of the TUBB6 (ENST00000591909) TSS.
}
\keyword{datasets}
