% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_documentation.R
\docType{data}
\name{tubb6_transcript_counts}
\alias{tubb6_transcript_counts}
\title{tubb6_transcript_counts}
\format{
A data.frame with normalized transcript counts for TUBB6 in 126 normal prostate samples.
}
\source{
RNA-seq data from 'Li, Jing, et al. "A genomic and epigenomic atlas of prostate cancer in Asian populations."
Nature 580.7801 (2020): 93-99.'
}
\usage{
tubb6_transcript_counts
}
\description{
Transcript counts for TUBB6 in normal prostate samples.
}
\keyword{datasets}
