% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tubb6_example_objects_documentation.R
\docType{data}
\name{tubb6_tss}
\alias{tubb6_tss}
\title{tubb6_tss}
\format{
GRanges object with 1 ranges for the TUBB6 TSS.
}
\source{
The TSS of the ENST00000591909 transcript.
}
\usage{
tubb6_tss
}
\description{
The location of the TSS for TUBB6.
}
\keyword{datasets}
