% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_documentation.R
\docType{data}
\name{TumourMethDatasets}
\alias{TumourMethDatasets}
\title{TumourMethDatasets}
\format{
A data.frame with one row for each dataset
}
\usage{
TumourMethDatasets
}
\description{
A table describing the datasets available from TumourMethData.
}
\keyword{datasets}
