% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_region_methylation.R
\name{.summarize_chunk_methylation}
\alias{.summarize_chunk_methylation}
\title{Summarize methylation values for regions in a chunk}
\usage{
.summarize_chunk_methylation(
  chunk_regions,
  meth_rse,
  assay_number,
  summary_function
)
}
\arguments{
\item{chunk_regions}{Chunk with genomic regions of interest.}

\item{meth_rse}{A RangedSummarizedExperiment with methylation values.}

\item{assay_number}{The assay from meth_rse to extract values from.}

\item{summary_function}{summary_function A function that summarizes column values.}
}
\value{
A function which returns a list with the
}
\description{
Summarize methylation values for regions in a chunk
}
