% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_cluster.R
\name{.tss_iterator}
\alias{.tss_iterator}
\title{Create an iterator function for use with bpiterate}
\usage{
.tss_iterator(
  meth_values_chunk,
  tss_region_indices_list,
  transcript_values,
  tss_for_chunk
)
}
\arguments{
\item{tss_for_chunk}{A list of GRanges with the TSS for the current chunk}
}
\value{
A function which returns a list with the
}
\description{
Create an iterator function for use with bpiterate
}
