% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_meth_site_transcript_cors.R,
%   R/setup_cluster.R
\name{.tss_iterator}
\alias{.tss_iterator}
\title{Create an iterator function for use with bpiterate}
\usage{
.tss_iterator(
  meth_values_chunk,
  tss_region_indices_list,
  transcript_values,
  tss_for_chunk
)

.tss_iterator(
  meth_values_chunk,
  tss_region_indices_list,
  transcript_values,
  tss_for_chunk
)
}
\arguments{
\item{meth_values_chunk}{A table with methylation values for current chunk}

\item{tss_region_indices_list}{A list with the indices for methylation sites associated with each TSS.}

\item{transcript_values}{A list with expression values for transcripts.}

\item{tss_for_chunk}{A list of GRanges with the TSS for the current chunk}
}
\value{
A function which returns a list with the

A function which returns a list with the
}
\description{
Create an iterator function for use with bpiterate

Create an iterator function for use with bpiterate
}
