% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{expand_granges}
\alias{expand_granges}
\title{Expand GRanges.}
\usage{
expand_granges(genomic_regions, upstream = 0, downstream = 0)
}
\arguments{
\item{genomic_regions}{A GRanges object}

\item{upstream}{Number of bases to add upstream of each region in genomic_regions. Must be numeric vector of length 1 or else equal to the length of genomic_regions. Default is 0.}

\item{downstream}{Number of bases to add downstream of each region in genomic_regions. Must be numeric vector of length 1 or else equal to the length of genomic_regions. Default is 0.}
}
\value{
None
}
\description{
Expand ranges in a GRanges object upstream and downstream by specified numbers of bases, taking strand into account.
If GRanges do not have a strand, they are treated like they are on the "+" strand.
}
