% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/granges_manipulation.R
\name{extractMethSitesFromGenome}
\alias{extractMethSitesFromGenome}
\title{Create a GRanges with methylation sites of interest from a BSgenome or DNAStringSet}
\usage{
extractMethSitesFromGenome(
  genome,
  pattern = "CG",
  stranded = TRUE,
  standard_sequences_only = TRUE
)
}
\arguments{
\item{genome}{A BSgenome object or a DNAStringSet with names indicating the sequences.}

\item{pattern}{A pattern to match in genome. Default is "CG".}

\item{stranded}{TRUE or FALSE indicating whether to return matches on
both strands or else just the "+" strand. Strand will be set to "*" if FALSE. Default is TRUE.}

\item{standard_sequences_only}{TRUE or FALSE indicating whether to only return ranges
on standard sequences (those without "_" in their names). Default is TRUE.}
}
\value{
A GRanges object with genomic regions matching the pattern.
}
\description{
Create a GRanges with methylation sites of interest from a BSgenome or DNAStringSet
}
\examples{
# Get human CpG sites for hg38 genome build
hg38_cpgs <- methodical::extractMethSitesFromGenome("BSgenome.Hsapiens.UCSC.hg38")
head(hg38_cpgs)

# Find CHG sites in Arabidopsis thaliana
arabidopsis_CHG_sites <- methodical::extractMethSitesFromGenome("BSgenome.Athaliana.TAIR.TAIR9", pattern = "CHG")
head(arabidopsis_CHG_sites)
}
