% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/granges_accessory_functions.R
\name{extractMethSitesFromGenome}
\alias{extractMethSitesFromGenome}
\title{Create a GRanges with methylation sites of interest from a BSgenome or DNAStringSet}
\usage{
extractMethSitesFromGenome(
  genome,
  pattern = "CG",
  stranded = TRUE,
  standard_seqs_only = FALSE
)
}
\arguments{
\item{genome}{A BSgenome object or a DNAStringSet with names indicating the sequences.}

\item{pattern}{A pattern to match in genome. Default is "CG".}

\item{stranded}{TRUE or FALSE indicating whether to return matches on
both strands or else just the "+" strand. Strand will be set to "*" if FALSE. Default is TRUE.}

\item{standard_seqs_only}{TRUE or FALSE indicating whether to only return ranges
on standard sequences (those without "_" in their names). Default is FALSE.}
}
\value{
A GRanges object with genomic regions matching the pattern.
}
\description{
Create a GRanges with methylation sites of interest from a BSgenome or DNAStringSet
}
\examples{
# Get human CpG sites for a portion of chr1 from hg38 genome build
data(chr1_subset_hg38, package = "methodical")
chr1_subset_hg38_cpgs <- methodical::extractMethSitesFromGenome(chr1_subset_hg38)
head(chr1_subset_hg38_cpgs)

# Find CHG sites in Arabidopsis thaliana
data(chr4_subset_a_thal, package = "methodical")
chr4_subset_a_thal_chg_sites <- methodical::extractMethSitesFromGenome(
genome = chr4_subset_a_thal, pattern = "CHG")
head(chr4_subset_a_thal_chg_sites)
}
