% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kallisto_quantify.R
\name{kallisto_index}
\alias{kallisto_index}
\title{Create an index file for running Kallisto}
\usage{
kallisto_index(
  path_to_kallisto,
  transcripts_fasta,
  index_name = "kallisto_index.idx"
)
}
\arguments{
\item{path_to_kallisto}{Path to kallisto executable}

\item{transcripts_fasta}{Path to a fasta file for the transcripts to be quantified.}

\item{index_name}{Name to give the created index file. Default is "kallisto_index.idx".}
}
\description{
Create an index file for running Kallisto
}
