% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_meth_rse.R
\name{makeMethRSEFromInputFiles}
\alias{makeMethRSEFromInputFiles}
\title{Create a HDF5-backed RangedSummarizedExperiment for methylation values in meth_files}
\usage{
makeMethRSEFromInputFiles(
  meth_files,
  seqnames_col,
  start_col,
  total_reads_col = NULL,
  meth_reads_col = NULL,
  unmeth_reads_col = NULL,
  meth_fraction_col = NULL,
  zero_based,
  meth_sites,
  sequence_context = "CG",
  collapse_strands = TRUE,
  decimal_places = NA,
  sample_metadata = NULL,
  hdf5_dir,
  overwrite = FALSE,
  chunkdim = NULL,
  temporary_dir = NULL,
  BPPARAM = BiocParallel::SerialParam(),
  ...
)
}
\arguments{
\item{meth_files}{A vector of paths to input methylation files. All sites in each file are assumed to be for the
same sequence context e.g. CG or CHG. Automatically detects if meth_files contain a header if every field in the first line is a character.}

\item{seqnames_col}{The column number in meth_files which corresponds to the sequence names.}

\item{start_col}{The column number in meth_files which corresponds to the genomic start coordinate.}

\item{total_reads_col}{The column number in meth_files which corresponds to the total number of reads for the position.}

\item{meth_reads_col}{The column number in meth_files which corresponds to the number of methylated reads for the position.}

\item{unmeth_reads_col}{The column number in meth_files which corresponds to the number of unmethylated reads for the position.}

\item{meth_fraction_col}{The column number in meth_files which corresponds to the fraction of reads that support methylation at the position.
Will be converted to a proportion if it appears to be a fraction.}

\item{zero_based}{TRUE or FALSE indicating if files are zero-based.}

\item{meth_sites}{A GRanges object with non-overlapping locations of methylation sites of interest e.g. CpG sites.
Any methylation sites in meth_files that are not in meth_sites are ignored.}

\item{sequence_context}{A single character string or DNAString with the
sequence context of the methylation sites e.g. CG or CHG. If a character, must be coercible to a DNAString. Default is "CG".}

\item{collapse_strands}{TRUE or FALSE indicating whether or not to collapse data on + and - strands.
Only makes sense for symmetrically methylated contexts e.g. CG or CHG and meth_sites should include ranges on both the + and - strands if TRUE.}

\item{decimal_places}{Optional integer indicating the number of decimal places to round beta values to. Default is not to round.}

\item{sample_metadata}{Sample metadata to be used as colData for the RangedSummarizedExperiment.}

\item{hdf5_dir}{Directory to save HDF5 file. Is created if it doesn't exist. HDF5 file is called assays.h5.}

\item{overwrite}{TRUE or FALSE indicating whether to allow overwriting if hdf5_dir already exists. Default is FALSE.}

\item{chunkdim}{The dimensions of the chunks for the HDF5 file. Should be a vector of length 2 giving the number of rows and then the number of columns in each chunk.
Uses HDF5Array::getHDF5DumpChunkDim(length(meth_sites), length(meth_files))) by default.}

\item{temporary_dir}{Name to give temporary directory created to store intermediate files. A directory with this name cannot already exist.
Default is to create a subdirectory named "temporary_meth_chunks_" inside the directory given by \code{tempdir()}. Will be deleted after completion.}

\item{BPPARAM}{A BiocParallelParam object for parallel processing. Defaults to \code{BiocParallel::SerialParam()}.}

\item{...}{Additional arguments to be passed to HDF5Array::HDF5RealizationSink() for controlling the physical properties of the created HDF5 file,
such as compression level. Uses the defaults for any properties that are not specified.}
}
\value{
A RangedSummarizedExperiment with two assays for all methylation sites in meth_sites,
beta with the proportion of methylated reads and Cov with the total  number of reads for each site. methylation sites will be in the same order as sort(meth_sites).
}
\description{
Create a HDF5-backed RangedSummarizedExperiment for methylation values in meth_files
}
\examples{

# Load CpGs within first million base pairs of chromosome 1 as a GRanges object
data("hg38_cpgs_subset", package = "methodical")

# Get paths to meth_files
meth_files <- list.files(path = system.file('extdata', package = 'methodical'), 
  pattern = ".bg.gz", full.names = TRUE)

# Create sample metadata
sample_metadata <- data.frame(
  tcga_project = gsub("_.*", "", gsub("TCGA_", "", basename(meth_files))),
  sample_type = ifelse(grepl("N", basename(meth_files)), "Normal", "Tumour"),
  row.names = tools::file_path_sans_ext(basename(meth_files))
)

# Create a HDF5-backed RangedSummarizedExperiment from meth_files
meth_rse <- makeMethRSEFromInputFiles(meth_files = meth_files, 
  seqnames_col = 1, start_col = 2, meth_reads_col = 4, unmeth_reads_col = 5, 
  zero_based = TRUE, meth_sites = hg38_cpgs_subset, sample_metadata = sample_metadata, 
  hdf5_dir = paste0(tempdir(), "/bedgraph_hdf5_1"))
  
}
