% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_meth_rse.R
\name{methrixToRSE}
\alias{methrixToRSE}
\title{Convert a Methrix object into a RangedSummarizedExperiment}
\usage{
methrixToRSE(methrix, assays = c("beta", "cov"))
}
\arguments{
\item{methrix}{A methrix object}

\item{assays}{A vector indicating the names of assays in methrix used to create a RangedSummarizedExperiment. Can be one or both of "beta" and "cov".
Default is both "beta" and "cov" assays.}
}
\value{
A RangedSummarizedExperiment
}
\description{
Convert a Methrix object into a RangedSummarizedExperiment
}
\examples{
# Load a sample methrix object
data("methrix_data", package = "methrix")
  
# Convert methrix to a RangedSummarizedExperiment with one assay for the methylation beta values
meth_rse <- methodical::methrixToRSE(methrix_data, assays = "beta")

}
