% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_meth_site_values}
\alias{plot_meth_site_values}
\title{Create plot of methylation values}
\usage{
plot_meth_site_values(
  meth_site_values,
  column_name,
  reference_region = NULL,
  title = NULL,
  xlabel = "Genomic Position",
  ylabel = "Methylation Site Values",
  low_colour = "cadetblue4",
  high_colour = "firebrick3"
)
}
\arguments{
\item{meth_site_values}{A data.frame with values associated with methylation sites.
First column should be called meth_site and give the names of methylation sites.
All methylation sites must be located on the same sequence.}

\item{column_name}{Name of column in meth_site_values to plot.}

\item{reference_region}{An optional GRanges object with a single range. If provided, the x-axis will
show the distance of methylation sites to the start of this region with methylation sites upstream
relative to the reference_region shown first. If not, the x-axis will show the start site coordinate of the methylation site.}

\item{title}{Title of the plot. Default is no title.}

\item{xlabel}{Label for the X axis in the plot. Default is "Genomic Position".}

\item{ylabel}{Label for the Y axis in the plot. Default is "Methylation Site Values".}

\item{low_colour}{Colour to use for low values. Default value is "cadetblue4".}

\item{high_colour}{Colour to use for high values. Default value is "firebrick3".}
}
\value{
A ggplot object
}
\description{
Create plot of methylation values
}
\examples{
# Load methylation-transcript correlation results for TUBB6 gene
data("tubb6_cpg_meth_transcript_cors", package = "methodical")

# Plot methylation-transcript correlation values around TUBB6 TSS
methodical::plot_meth_site_values(tubb6_cpg_meth_transcript_cors, column_name = "cor", ylabel = "Spearman Correlation")

# Create same plot but showing the distance to the TUBB6 TSS on the x-axis
methodical::plot_meth_site_values(tubb6_cpg_meth_transcript_cors, column_name = "cor", 
  ylabel = "Spearman Correlation", reference_region = attributes(tubb6_cpg_meth_transcript_cors)$tss_range)

}
