% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_documentation.R
\docType{data}
\name{tubb6_meth_rse}
\alias{tubb6_meth_rse}
\title{tubb6_meth_rse}
\format{
A call to create a RangedSummarizedExperiment with methylation data for 355 CpG sites within +/- 5,000
base pairs of the TUBB6 TSS in 126 normal prostate samples.
Should be evaluated after loading using \code{tubb6_meth_rse <- tubb6_meth_rse <- eval(tubb6_meth_rse)} to restore the RangedSummarizedExperiment.
}
\source{
WGBS data from 'Li, Jing, et al. "A genomic and epigenomic atlas of prostate cancer in Asian populations."
Nature 580.7801 (2020): 93-99.'
}
\usage{
tubb6_meth_rse
}
\description{
The location of the TSS for TUBB6.
}
\keyword{datasets}
